src/requirements.txt
python-telegram-bot==13.7
requests==2.31.0
yfinance==0.2.40